pyproject.toml
[tool.poetry]
name = "whenx"
version = "0.1.0"
description = ""
authors = ["Edmar Ferreira <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.15"
mysql-connector-python = "^8.0.33"
python-dotenv = "^1.0.0"
openai = "^0.27.7"
google-search-results = "^2.4.2"
faiss-cpu = "^1.7.4"
tiktoken = "^0.4.0"
bs4 = "^0.0.1"
langchain = "^0.0.200"
playwright = "^1.35.0"
alembic = "^1.11.1"
schedule = "^1.2.0"
resend = "^0.5.1"
prettytable = "^3.8.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"
pytest = "^7.3.1"
ipykernel = "^6.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
